pyproject.toml
[project]
name = "doc-extractor"
version = "0.1.0"
description = "Extract structured JSON from supply chain documents using Google Gemini/Gemma models"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai>=1.0.0",
    "pydantic>=2.0.0",
    "liteparse",
    "datalab-python-sdk>=0.5.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "rapidfuzz>=3.14.5",
]

[tool.pytest.ini_options]
pythonpath = ["scripts"]
testpaths = ["tests"]
